requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py37h7deecbd_0
brotlipy=0.7.0=py37h540881e_1004
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
cairo=1.16.0=h19f5f5c_2
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.14.6=py37hc58025e_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
cryptography=37.0.2=py37h38fbfac_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py37h06a4308_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.0.0=hf484d3e_5
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jasper=2.0.14=hd8c5072_2
jedi=0.18.1=py37h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.1=py37h06a4308_0
jpeg=9e=h7f8727e_0
jsonschema=2.6.0=py37_0
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
matplotlib-inline=0.1.6=py37h06a4308_0
mistune=0.8.4=py37h14c3975_1001
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
nbconvert=5.5.0=py_0
nbformat=5.5.0=py37h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py37h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
notebook=5.7.10=py37_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.3=py37h807cd23_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1v=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h8c16a72_0
pandoc=2.12=h06a4308_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle5=0.0.11=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py37hace64e9_1
pip=22.2.2=py37h06a4308_0
pixman=0.40.0=h7f8727e_1
ply=3.11=py37_0
prometheus_client=0.14.1=py37h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
psutil=5.9.0=py37h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pysocks=1.7.1=py37h89c1867_5
python=3.7.13=haa1d7c7_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.12.1=py3.7_cuda11.6_cudnn8.3.2_0
pytorch-fid=0.3.0=pypi_0
pytorch-msssim=0.2.1=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py37h06a4308_0
pyyaml=6.0=py37h5eee18b_1
pyzmq=23.2.0=py37h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.3.2=py37h06a4308_0
qtpy=2.2.0=py37h06a4308_0
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.2=pyhd8ed1ab_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37hc147768_0
seaborn=0.11.0=py_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py37h06a4308_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
terminado=0.13.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.12.1=py37_cu116
torchvision=0.13.1=cpu_py37h164cc8f_0
tornado=6.2=py37h5eee18b_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.14=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h06a4308_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
